create-2fa-secret-and-qrcode/requirements.txt
psycopg2-binary
cryptography
pyotp
qrcode
pillow
